requirements.txt
asgiref==3.2.7
astroid==2.9.3
beautifulsoup4==4.12.3
black==23.11.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
Django==3.0.5
django-cors-headers==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.15.1
exceptiongroup==1.2.0
flake8==6.1.0
fonttools==4.47.2
idna==3.7
iniconfig==2.0.0
isort==5.10.1
kiwisolver==1.4.5
lazy-object-proxy==1.7.1
linkedlist==2.0.0
matplotlib==3.8.2
mccabe==0.7.0
mypy-extensions==1.0.0
node==0.9.28
numpy==1.26.3
odict==1.8.0
packaging==23.2
pandas==2.1.4
pathspec==0.11.2
pillow==10.2.0
platformdirs==2.4.1
pluggy==1.4.0
plumber==1.6
pycodestyle==2.11.0
pycparser==2.22
pyflakes==3.1.0
pylint==2.12.2
PyMySQL==1.1.0
pyparsing==3.1.1
pytest==8.0.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2020.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
six==1.16.0
soupsieve==2.5
sqlparse==0.3.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.2.1
wikipedia==1.4.0
wrapt==1.13.3
zope.component==5.0.1
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.1.0
zope.interface==5.4.0
zope.lifecycleevent==4.3
